Diary -> HomeNavigator 변경
